bootloader/lkenv: mv SnapBootSelect_v1 to separate file
This will simplify the diff for future commits.
